backend_python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.6.0
dspy-ai==2.4.0
ffmpeg-python==0.2.0
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0
google-genai
